mcp-backend/pyproject.toml
[project]
name = "mcp-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "mcp>=1.13.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
